[FIX] try to fix some bugs

+++ src/test/groovy/io/choerodon/devops/api/controller/v1/CiControllerSpec.groovy
@@ -146,34 +146,34 @@ class CiControllerSpec extends Specification {
 
         then: '校验文件是否创建'
         File gzFile = new File("/Charts/org/pro/key.tar.gz")
-        gzFile.getName()=="key.tar.gz"
+        gzFile.getName() == "key.tar.gz"
         File yamlFile = new File("/devopsversion/values.yaml")
         yamlFile.getName() == "values.yaml"
 
         // 删除app
         List<ApplicationDO> list = applicationMapper.selectAll()
-        if (!list.isEmpty()) {
+        if (list != null && !list.isEmpty()) {
             for (ApplicationDO e : list) {
                 applicationMapper.delete(e)
             }
         }
         // 删除appVersion
         List<ApplicationVersionDO> list1 = applicationVersionMapper.selectAll()
-        if (!list1.isEmpty()) {
+        if (list1 != null && !list1.isEmpty()) {
             for (ApplicationVersionDO e : list1) {
                 applicationVersionMapper.delete(e)
             }
         }
         // 删除appVersionReadme
         List<ApplicationVersionReadmeDO> list2 = applicationVersionReadmeMapper.selectAll()
-        if (!list2.isEmpty()) {
+        if (list2 != null && !list2.isEmpty()) {
             for (ApplicationVersionReadmeDO e : list2) {
                 applicationVersionReadmeMapper.delete(e)
             }
         }
         // 删除appVersionValue
         List<ApplicationVersionValueDO> list3 = applicationVersionValueMapper.selectAll()
-        if (!list3.isEmpty()) {
+        if (list3 != null && !list3.isEmpty()) {
             for (ApplicationVersionValueDO e : list3) {
                 applicationVersionValueMapper.delete(e)
             }
